requirements.txt
streamlit
scikit-learn
pandas
numpy
streamlit
streamlit
openai
langchain
tqdm
matplotlib
seaborn
python-dotenv
transformers
huggingface-hub
torch
plotly


